Fixed bug in pairwiseLCS
If the lineage of taxon A was shorter than the lineage of taxon B not the LCA was calculated, but a taxon more basal in the tree.
This bug was identified by preparing test cases.

Moreover I added a test for same lineage of A and B. In this case, one can return the lineage immediately